active_interview_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.10.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coverage==7.6.12
cryptography==44.0.1
cssbeautifier==1.15.4
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.19
django-bootstrap-v5==1.0.11
django-filter==25.1
djangorestframework==3.15.2
djlint==1.36.4
docx==0.2.4
EditorConfig==0.17.0
filetype==1.2.0
flake8==7.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.11.1
jsbeautifier==1.15.4
json5==0.10.0
lxml==5.4.0
Markdown==3.7
markdownify==1.1.0
mccabe==0.7.0
openai==2.5.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.3.2
PyMuPDF==1.25.5
pymupdf4llm==0.0.20
PySocks==1.7.1
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
scapy==2.6.1
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
style==1.1.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
update==0.0.1
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.6.0
wsproto==1.2.0
django-allauth==0.57.0
django-ratelimit==4.1.0
pytest==8.3.4
pytest-django==4.9.0
icalendar==6.1.0

